+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import Myrouter from './components/Myrouter';
+import Myrouter from './Pages/Myrouter';
 
 function App() {
   return (
