+++ src/Pages/About.jsx
@@ -1,8 +1,10 @@
 import React from 'react'
-
+import Card from '../components/Card'
 const About = () => {
   return (
-    <div>About</div>
+    <div>
+      <Card/>
+    </div>
   )
 }
