+++ src/components/Card/Card.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Card = () => {
+  return (
+    <div>Card</div>
+  )
+}
+
+export default Card
