+++ src/components/Card/index.js
@@ -1,0 +1,3 @@
+
+import Card from "./Card";
+export default Card
